+++ src/main/java/io/github/mzmine/datamodel/features/types/annotations/GNPSLibraryUrlType.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2006-2020 The MZmine Development Team
+ *
+ * This file is part of MZmine.
+ *
+ * MZmine is free software; you can redistribute it and/or modify it under the terms of the GNU
+ * General Public License as published by the Free Software Foundation; either version 2 of the
+ * License, or (at your option) any later version.
+ *
+ * MZmine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
+ * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+ * Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with MZmine; if not,
+ * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
+ */
+
+package io.github.mzmine.datamodel.features.types.annotations;
+
+import io.github.mzmine.datamodel.features.types.abstr.UrlType;
+import org.jetbrains.annotations.NotNull;
+
+/**
+ * Link to a GNPS library entry
+ *
+ * @author Robin Schmid (https://github.com/robinschmid)
+ */
+public class GNPSLibraryUrlType extends UrlType {
+
+  @NotNull
+  @Override
+  public String getHeaderString() {
+    return "GNPS library";
+  }
+}
