--- src/main/java/io/github/mzmine/datamodel/features/types/numbers/FormulaMassType.java
@@ -1,53 +1,0 @@
-/*
- * Copyright 2006-2020 The MZmine Development Team
- *
- * This file is part of MZmine.
- *
- * MZmine is free software; you can redistribute it and/or modify it under the terms of the GNU
- * General Public License as published by the Free Software Foundation; either version 2 of the
- * License, or (at your option) any later version.
- *
- * MZmine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
- * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
- * Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with MZmine; if not,
- * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
- * USA
- */
-
-package io.github.mzmine.datamodel.features.types.numbers;
-
-import io.github.mzmine.datamodel.features.types.annotations.FormulaType;
-import io.github.mzmine.datamodel.features.types.numbers.abstr.DoubleType;
-import io.github.mzmine.main.MZmineCore;
-import io.github.mzmine.util.FormulaUtils;
-import java.text.DecimalFormat;
-import java.text.NumberFormat;
-
-/**
- * The formula mass typically follows a {@link FormulaType} and describes the neutral mass or m/z
- * depending on the formula. Also see {@link FormulaUtils}.
- */
-public class FormulaMassType extends DoubleType {
-
-  public FormulaMassType() {
-    super(new DecimalFormat("0.0000"));
-  }
-
-  @Override
-  public NumberFormat getFormatter() {
-    try {
-      return MZmineCore.getConfiguration().getMZFormat();
-    } catch (NullPointerException e) {
-      // only happens if types are used without initializing the MZmineCore
-      return DEFAULT_FORMAT;
-    }
-  }
-
-  @Override
-  public String getHeaderString() {
-    return "Formula mass";
-  }
-
-}
